moved cellprofiler.utilities.hdf5_dict into the cellprofiler package (resolves #1933)
