+++ cellprofiler/HDF5.py
