+++ cellprofiler/measurement.py
@@ -8,21 +8,18 @@ import logging
 import h5py
 
 logger = logging.getLogger(__name__)
-import numpy as np
 import re
 from scipy.io.matlab import loadmat
-from itertools import repeat
 import cellprofiler.preferences as cpprefs
-from cellprofiler.utilities.hdf5_dict import HDF5Dict, get_top_level_group
-from cellprofiler.utilities.hdf5_dict import VERSION, HDFCSV, VStringArray
-from cellprofiler.utilities.hdf5_dict import HDF5ObjectSet
-from cellprofiler.utilities.hdf5_dict import NullLock
+from cellprofiler.HDF5 import HDF5Dict, get_top_level_group
+from cellprofiler.HDF5 import VERSION, HDFCSV, VStringArray
+from cellprofiler.HDF5 import HDF5ObjectSet
+from cellprofiler.HDF5 import NullLock
 import tempfile
 import numpy as np
 import warnings
 import os
 import os.path
-import mmap
 import urllib
 import sys
